requirements.txt
asgiref==3.7.0
Django==4.2.1
django-crispy-forms==2.0
django-filter==23.2
django-guardian==2.4.0
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.0